LayoutTests/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A5_T1-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Label 'LABEL_ANOTHER_LOOP' is not defined
S12.7_A5_T1

PASS Expected parsing failure

TEST COMPLETE
